requirements.txt
fastapi
sqlmodel
sqlalchemy
uvicorn
bcrypt
pyjwt
python-dotenv
'pydantic[email]'
python-multipart
slowapi